requirements.txt
Django==2.0
django-annoying==0.10.4
django-cors-headers==2.1.0
django-enumchoicefield==0.9.0
djangorestframework==3.7.4
gunicorn==19.7.1
mysqlclient==1.3.12
pytz==2017.3
six==1.11.0
